+++ python/GafferUSDUI/__init__.py
@@ -34,4 +34,6 @@
 #
 ##########################################################################
 
+from . import USDLayerWriterUI
+
 __import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", subdirectory = "GafferUSDUI" )